Update predatordiversity.md
Mary's comments, part 1. Great work Andrew, this is a fine project and I LOVE that you prepared this in github R markdown. My only real critique at this point is to dig a little deeper and take this thing by the horns conceptually. Make the language more active and strongly centered on the conceptual premise and advance of the work. Keep up the good work!